Cleans up metrics pkg to not require locks, and split out resources into different files
